+++ openeogeotrellis/deploy/batch_job.py
@@ -14,6 +14,7 @@ from copy import deepcopy
 from math import isfinite
 
 import pyproj
+from openeo_driver.workspacerepository import BackendConfigWorkspaceRepository, backend_config_workspace_repository
 from osgeo import gdal
 from py4j.protocol import Py4JJavaError
 from pyspark import SparkContext, SparkConf
@@ -1160,6 +1161,8 @@ def run_job(
 
         unique_process_ids = CollectUniqueProcessIdsVisitor().accept_process_graph(process_graph).process_ids
 
+        # this is subtle: result now points to the last of possibly several results (#295); it corresponds to
+        # the terminal save_result node of the process graph
         if "file_metadata" in result.options:
             result.options["file_metadata"]["providers"] = [
                 {
@@ -1182,6 +1185,8 @@ def run_job(
                                                     unique_process_ids=unique_process_ids,
                                                     asset_metadata=assets_metadata,
                                                     ml_model_metadata=ml_model_metadata)
+
+        _export_workspace(result, result_metadata, stac_metadata_dir=job_dir)
     finally:
         metadata = {**result_metadata, **_get_tracker_metadata("")}
 
@@ -1206,6 +1211,72 @@ def run_job(
                 s3_instance.upload_file(full_path, bucket, full_path.strip("/"))
 
         get_jvm().com.azavea.gdal.GDALWarp.deinit()
+
+
+def _export_workspace(result: SaveResult, result_metadata: dict, stac_metadata_dir: Path):
+    asset_paths = [Path(asset["href"]) for asset in result_metadata["assets"].values()]
+    stac_paths = _write_exported_stac_collection(stac_metadata_dir, result_metadata)
+    result.export_workspace(workspace_repository=backend_config_workspace_repository,
+                            files=asset_paths + stac_paths,
+                            default_merge=OPENEO_BATCH_JOB_ID)
+
+
+def _write_exported_stac_collection(job_dir: Path, result_metadata: dict) -> List[Path]:  # TODO: change to Set?
+    def write_stac_item_file(asset_id: str, asset: dict) -> Path:
+        item_file = job_dir / f"{asset_id}.json"
+
+        stac_item = {
+            "type": "Feature",
+            "stac_version": "1.0.0",
+            "id": asset_id,
+            "geometry": asset.get("geometry"),
+            "bbox": asset.get("bbox"),
+            "properties": {
+                "datetime": asset.get("datetime"),
+            },
+            "links": [],  # TODO
+            "assets": {
+                asset_id: dict_no_none(**{
+                    "href": f"./{Path(asset['href']).name}",
+                    "roles": asset.get("roles"),
+                    "type": asset.get("type"),
+                    "eo:bands": asset.get("bands"),
+                })
+            },
+        }
+
+        with open(item_file, "wt") as fi:
+            json.dump(stac_item, fi)
+
+        return item_file
+
+    item_files = [write_stac_item_file(asset_id, asset) for asset_id, asset in result_metadata["assets"].items()]
+
+    def item_link(item_file: Path) -> dict:
+        return {
+            "href": f"./{item_file.name}",
+            "rel": "item",
+            "type": "application/geo+json",
+        }
+
+    stac_collection = {
+        "type": "Collection",
+        "stac_version": "1.0.0",
+        "id": OPENEO_BATCH_JOB_ID,
+        "description": "TODO",  # TODO
+        "license": "TODO",  # TODO
+        "extent": {
+            "spatial": {"bbox": [[-180, -90, 180, 90]]},  # TODO
+            "temporal": {"interval": [[None, None]]}  # TODO
+        },
+        "links": [item_link(item_file) for item_file in item_files],
+    }
+
+    collection_file = job_dir / "collection.json"
+    with open(collection_file, "wt") as fc:
+        json.dump(stac_collection, fc)
+
+    return item_files + [collection_file]
 
 
 def _convert_job_metadatafile_outputs_to_s3_urls(metadata_file: Path):